+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.29.1...master
 
+### Fixed
+- Fix remote JVM debugger setup for Jira. Aid DCPERF-432.
+
 ## [4.29.1] - 2024-06-17
 [4.29.1]: https://github.com/atlassian/infrastructure/compare/release-4.29.0...release-4.29.1